eeprom-programmer/requirements.txt
pyserial
click
tqdm
